hoisted stickings props to top level, added outlet context props type
