+++ src/pages/MyAccount/MyAccount.tsx
@@ -1,13 +1,11 @@
-// import { useNavigate } from 'react-router-dom';
-import { useAuth } from '../../context/authContext/useAuth';
 import Login from './Login/Login';
 import UserAccount from './UserAccount/UserAccount';
+import { useAuth } from '../../context/authContext/useAuth';
 import { CircularProgress } from '@mui/material';
 import './MyAccount.css';
 
 const MyAccount = () => {
   const { authUser, isLoading } = useAuth();
-  // const navigate = useNavigate();
 
   return isLoading ? (
     <CircularProgress
